+++ tests/geant/test_geant3id.py
@@ -5,39 +5,39 @@
 
 import pytest
 
-from particle.geant import GeantID
+from particle.geant import Geant3ID
 from particle.pdgid import PDGID
 from particle.exceptions import MatchingIDNotFound
 
 
 def test_class_string_representations():
-    pid = GeantID(1)
+    pid = Geant3ID(1)
     assert pid == 1
-    assert pid.__str__() == "<GeantID: 1>"
+    assert pid.__str__() == "<Geant3ID: 1>"
 
 
 def test_class_return_type():
-    assert isinstance(-GeantID(3), GeantID)
-    assert isinstance(~GeantID(3), GeantID)
+    assert isinstance(-Geant3ID(3), Geant3ID)
+    assert isinstance(~Geant3ID(3), Geant3ID)
 
 
 def test_class_inversion():
-    assert -GeantID(1) == ~GeantID(1)
+    assert -Geant3ID(1) == ~Geant3ID(1)
 
 
 def test_from_pdgid():
-    assert GeantID.from_pdgid(211) == 8
+    assert Geant3ID.from_pdgid(211) == 8
 
-    assert GeantID.from_pdgid(PDGID(211)) == 8
-    assert GeantID.from_pdgid(PDGID(211)) == GeantID(8)
+    assert Geant3ID.from_pdgid(PDGID(211)) == 8
+    assert Geant3ID.from_pdgid(PDGID(211)) == Geant3ID(8)
 
 
 def test_from_pdgid_non_matching():
     with pytest.raises(MatchingIDNotFound):
-        gid = GeantID.from_pdgid(55)
+        gid = Geant3ID.from_pdgid(55)
 
 
 def test_to_pdgid():
-    geantid = GeantID(8)
-    assert geantid.to_pdgid() == 211
-    assert geantid.to_pdgid() == PDGID(211)
+    gid = Geant3ID(8)
+    assert gid.to_pdgid() == 211
+    assert gid.to_pdgid() == PDGID(211)