+++ docs/CHANGELOG.rst
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+Version 0.8.0
+-------------
+In preparation.
+
+* Changes in API:
+  - ``GeantID`` class renamed to ``Geant3ID``.
+* Data CSV files:
+  - Version 3 (and 2) of package data files.
+  - File ``data/pdgid_to_geantid.csv`` renamed to ``data/pdgid_to_geant3id.csv``.
+  - 2016 and 2017 PDG data files ``mass_width_2016.mcd`` and ``mass_width_2017.mcd`` removed.
+* Tests:
+  - Tests for ``converters`` submodule added.
+  - Tests for ``Geant3ID`` extended.
+* Miscellaneous:
+  - Package files formatted with ``Black`` package.
+
+
 Version 0.7.1
 -------------
 November 22nd, 2019